+++ tests/datetime_tests/test_add.py
@@ -145,42 +145,42 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 3, 31, 1, 59, 59, 999999)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 2013, 3, 31, 3, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(7200, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
         dt = pendulum.create(2013, 3, 10, 1, 59, 59, 999999, 'America/New_York')
 
         self.assertDateTime(dt, 2013, 3, 10, 1, 59, 59, 999999)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(-5 * 3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 2013, 3, 10, 3, 0, 0, 0)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(- 4 * 3600, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
         dt = pendulum.create(1957, 4, 28, 1, 59, 59, 999999, 'America/New_York')
 
         self.assertDateTime(dt, 1957, 4, 28, 1, 59, 59, 999999)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(-5 * 3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 1957, 4, 28, 3, 0, 0, 0)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(- 4 * 3600, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
     def test_add_time_to_new_transition_skipped_big(self):
         dt = pendulum.create(2013, 3, 31, 1, tz='Europe/Paris')
@@ -188,14 +188,14 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 3, 31, 1, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
         dt = dt.add(weeks=1)
 
         self.assertDateTime(dt, 2013, 4, 7, 1, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(7200, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
     def test_add_time_to_new_transition_repeated(self):
         dt = pendulum.create(2013, 10, 27, 1, 59, 59, 999999, 'Europe/Paris')
@@ -204,14 +204,14 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 10, 27, 2, 59, 59, 999999)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(7200, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 2013, 10, 27, 2, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
 
         dt = pendulum.create(2013, 11, 3, 0, 59, 59, 999999, 'America/New_York')
@@ -220,14 +220,14 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 11, 3, 1, 59, 59, 999999)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(-4 * 3600, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 2013, 11, 3, 1, 0, 0, 0)
         self.assertEqual('America/New_York', dt.timezone_name)
         self.assertEqual(-5 * 3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
     def test_add_time_to_new_transition_repeated_big(self):
         dt = pendulum.create(2013, 10, 27, 1, tz='Europe/Paris')
@@ -235,14 +235,14 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 10, 27, 1, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(7200, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())
 
         dt = dt.add(weeks=1)
 
         self.assertDateTime(dt, 2013, 11, 3, 1, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
     def test_add_time_to_new_transition_does_not_use_transition_rule(self):
         pendulum.set_transition_rule(pendulum.TRANSITION_ERROR)
@@ -251,11 +251,11 @@ class AddTest(AbstractTestCase):
         self.assertDateTime(dt, 2013, 3, 31, 1, 59, 59, 999999)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(3600, dt.offset)
-        self.assertFalse(dt.is_dst)
+        self.assertFalse(dt.is_dst())
 
         dt = dt.add(microseconds=1)
 
         self.assertDateTime(dt, 2013, 3, 31, 3, 0, 0, 0)
         self.assertEqual('Europe/Paris', dt.timezone_name)
         self.assertEqual(7200, dt.offset)
-        self.assertTrue(dt.is_dst)
+        self.assertTrue(dt.is_dst())