+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - `Interval` class has been renamed to `Duration`.
 - `strict` keyword argument for `parse` has been renamed to `exact`.
 - `at()` now supports setting partial time.
+- `local`, `utc` and `is_dst` are now methods rather than properties (`is_local()`, `is_utc()`, `is_dst()`).
 - Made the `strict` keyword argument for `parse` false by default, which means it will not fallback on the `dateutil` parser.
 - Improved performances of the `precise_diff()` helper.